+++ src/main/java/com/github/maracas/roseau/combinatorial/v2/benchmark/Benchmark.java
@@ -80,7 +80,8 @@ public final class Benchmark implements Runnable {
 			}
 		}
 
-		ExplorerUtils.removeDirectory(benchmarkWorkingPath);
+		if (errorsCount == 0)
+			ExplorerUtils.removeDirectory(benchmarkWorkingPath);
 	}
 
 	public void informsBreakingApisGenerationIsOver() {
